Refactored code and Unified Index
