+++ tools/issue-labeler/src/SearchIndexCreator/DocumentTypes.cs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace SearchIndexCreator
+{
+    public enum DocumentTypes
+    {
+        Issue,
+        Document
+    }
+}
